Reorganization and update settings
